# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JA4proxy Management UI
  description: Management interface for JA4proxy TLS security proxy
  version: 1.0.0
paths:
...
  /api/v1/events:
    get:
      tags:
      - events
      summary: Stream Events
      description: 'Stream live connection events as Server-Sent Events.


        Clients should connect with ``Accept: text/event-stream``.

        The stream never ends — clients must disconnect when done.


        When the client requests ``Accept: application/json`` (e.g. in tests or

        when probing the endpoint), a non-streaming JSON acknowledgement is returned

        instead of an infinite SSE stream.'
      operationId: stream_events_api_v1_events_get
      parameters:
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/webhooks:
    post:
      tags:
      - webhooks
      summary: Create Webhook
      description: 'Create a new webhook subscription.


        Returns the full record plus a one-time plaintext secret.

        The secret is NOT stored; only its bcrypt hash is persisted.'
      operationId: create_webhook_api_v1_webhooks_post
      parameters:
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - webhooks
      summary: List Webhooks
      description: Return all webhook subscriptions (no secrets).
      operationId: list_webhooks_api_v1_webhooks_get
      parameters:
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/webhooks/{webhook_id}:
    get:
      tags:
      - webhooks
      summary: Get Webhook
      description: Return a single webhook by ID (no secrets).
      operationId: get_webhook_api_v1_webhooks__webhook_id__get
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: string
          title: Webhook Id
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - webhooks
      summary: Update Webhook
      description: 'Update an existing webhook in-place.


        Only ``url``, ``events``, and ``active`` may be changed.

        Returns the updated resource.'
      operationId: update_webhook_api_v1_webhooks__webhook_id__put
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: string
          title: Webhook Id
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - webhooks
      summary: Delete Webhook
      description: 'Delete a webhook subscription.


        Idempotent — returns 204 whether or not the webhook existed.'
      operationId: delete_webhook_api_v1_webhooks__webhook_id__delete
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: string
          title: Webhook Id
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WebhookCreate:
      properties:
        url:
          type: string
          maxLength: 2048
          minLength: 1
          title: Url
        events:
          items:
            type: string
          type: array
          title: Events
        active:
          type: boolean
          title: Active
          default: true
      type: object
      required:
      - url
      title: WebhookCreate
      description: Body for POST /api/v1/webhooks.
    WebhookUpdate:
      properties:
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        events:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Events
        active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
      type: object
      title: WebhookUpdate
      description: Body for PUT /api/v1/webhooks/{id}.